split up types
